Initial debian scripts to generate .deb files
This is a set of files used to debianized paflib in order to generated
paflib deb files. This file was already sent to debian as an ITP and the
package is also at mentors.

With this script, three binary packages are generated, one for dsc library,
another one for ebb library and a -dev package.